Enable multiple dust template namespaces
There are times when it is useful to have different template
namespace. i.e: when template1 includes base, it may want to be
referring to a different base than template2. This is not currently
possible in dust as the cache is a global variable.

This patch provides the simplest possible modification of the
current implementation to enable multiple namespaces.

The library is changed to provide a single create method. This
method returns an object that provides all the functions that
are normally available on the dust module. Each object returned
from create uses its own cache variable, and hence provides its
own unique namespace.